+++ R/doy_noy.R
@@ -7,6 +7,7 @@
 #' @param lon numeric. Longitude in decimal degrees.
 #' @param lat numeric. Latitude in decimal degrees.
 #' @param method method by which to do the time zone lookup. Either \code{"fast"} (default) or \code{"accurate"}, see \link[lutz]{tz_lookup_coords}.
+#' @param ... optional lat,lon arguments.
 #'
 #' @name doy_noy
 #'
@@ -14,6 +15,16 @@
 #' First night of the year is the night with datetime Jan 01 00:00:00 in the local time zone,
 #' i.e. sunset on Jan 1 occurs on the second night of the year, and New Years Eve on Dec 31
 #' occurs on the first night of the new year.
+#'
+#' @examples
+#' # night of year of a profile (vp object):
+#' noy(example_vp)
+#'
+#' # day of year of a profile (vp object):
+#' noy(example_vp)
+#'
+#' # night of year of a vertial profile time series (vpts object):
+#' noy(example_vpts)
 NULL
 
 #' @rdname doy_noy
@@ -33,7 +44,7 @@ noy <- function(x, ..., method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-doy.default <- function(x, lon, lat, method = "fast") {
+doy.default <- function(x, lon, lat, ..., method = "fast") {
   tzone = lutz::tz_lookup_coords(lat, lon, method = method, warn = FALSE)
   yday(lubridate::with_tz(x, tzone = tzone))
 }
@@ -41,14 +52,14 @@ doy.default <- function(x, lon, lat, method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-noy.default <- function(x, lon, lat, method = "fast") {
+noy.default <- function(x, lon, lat, ..., method = "fast") {
   doy.default(x + 12 *3600, lon, lat, method = method)
 }
 
 #' @rdname doy_noy
 #'
 #' @export
-doy.vp <- function(x, method = "fast") {
+doy.vp <- function(x, ..., method = "fast") {
   stopifnot(inherits(x, "vp"))
   doy(x$datetime, x$attributes$where$lon, x$attributes$where$lat, method = method)
 }
@@ -56,7 +67,7 @@ doy.vp <- function(x, method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-noy.vp <- function(x, method = "fast") {
+noy.vp <- function(x, ..., method = "fast") {
   stopifnot(inherits(x, "vp"))
   noy(x$datetime, x$attributes$where$lon, x$attributes$where$lat, method = method)
 }
@@ -64,7 +75,7 @@ noy.vp <- function(x, method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-doy.vpts <- function(x, method = "fast") {
+doy.vpts <- function(x, ..., method = "fast") {
   stopifnot(inherits(x, "vpts"))
   doy(x$datetime, x$attributes$where$lon, x$attributes$where$lat, method = method)
 }
@@ -72,7 +83,7 @@ doy.vpts <- function(x, method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-noy.vpts <- function(x, method = "fast") {
+noy.vpts <- function(x, ..., method = "fast") {
   stopifnot(inherits(x, "vpts"))
   noy(x$datetime, x$attributes$where$lon, x$attributes$where$lat, method = method)
 }
@@ -80,7 +91,7 @@ noy.vpts <- function(x, method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-doy.vpi <- function(x, method = "fast") {
+doy.vpi <- function(x, ..., method = "fast") {
   stopifnot(inherits(x, "vpi"))
   doy(x$datetime, attributes(x)$lon, attributes(x)$lat, method = method)
 }
@@ -88,7 +99,7 @@ doy.vpi <- function(x, method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-noy.vpi <- function(x, method = "fast") {
+noy.vpi <- function(x, ..., method = "fast") {
   stopifnot(inherits(x, "vpi"))
   noy(x$datetime, attributes(x)$lon, attributes(x)$lat, method = method)
 }
@@ -96,7 +107,7 @@ noy.vpi <- function(x, method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-doy.pvol <- function(x, method = "fast") {
+doy.pvol <- function(x, ..., method = "fast") {
   stopifnot(inherits(x, "pvol"))
   doy(x$datetime, x$attributes$where$lon, x$attributes$where$lat, method = method)
 }
@@ -104,7 +115,7 @@ doy.pvol <- function(x, method = "fast") {
 #' @rdname doy_noy
 #'
 #' @export
-noy.pvol <- function(x, method = "fast") {
+noy.pvol <- function(x, ...,  method = "fast") {
   stopifnot(inherits(x, "pvol"))
   noy(x$datetime, x$attributes$where$lon, x$attributes$where$lat, method = method)
 }